Update and refactor Python files

+++ win_scripts/clean.py
@@ -10,6 +10,7 @@ import shutil
 import sys
 import subprocess
 
+
 def fail(msg):
     print(f"{sys.argv[0]}: {msg}", file=sys.stderr)
     sys.exit(111)
@@ -23,14 +24,17 @@ def try_run(command):
 
 
 def clean_files(directory):
+    deleted = False
     for filename in os.listdir(directory):
         file_path = os.path.join(directory, filename)
         if os.path.isfile(file_path):
             try:
                 os.remove(file_path)
+                deleted = True
                 print(f"Removed: {file_path}")
             except Exception as e:
                 fail(f"Failed to remove {file_path}: {e}")
+    return deleted
 
 
 def delete_directory(directory):
@@ -38,27 +42,55 @@ def delete_directory(directory):
         try:
             shutil.rmtree(directory)
             print(f"Removed directory: {directory}")
+            return True
         except Exception as e:
             fail(f"Failed to remove directory {directory}: {e}")
+    return False
 
 
 def display_help():
     print("\nScript to remove unneeded artifacts\n")
 
-if '--help' in sys.argv:
+
+if "--help" in sys.argv:
     display_help()
     sys.exit(0)
 
 
-# Set chromium/src dir from Windows environment variable
-cr_src_dir = os.getenv('CR_DIR', r'C:/src/chromium/src')
+def main():
+    cr_src_dir = os.getenv("CR_DIR", r"C:\src\chromium\src")
+    profiles_dir = os.path.normpath(
+        os.path.join(cr_src_dir, "chrome", "build", "pgo_profiles")
+    )
+    thorium_dir = os.path.normpath(os.path.join(cr_src_dir, "out", "thorium"))
 
-print("\nCleaning up unneeded artifacts\n")
+    cleanup_needed = any(
+        [
+            (
+                os.path.isdir(profiles_dir)
+                and any(
+                    os.path.isfile(os.path.join(profiles_dir, filename))
+                    for filename in os.listdir(profiles_dir)
+                )
+            ),
+            os.path.exists(thorium_dir),
+        ]
+    )
 
-profiles_dir = os.path.normpath(os.path.join(cr_src_dir, "chrome", "build", "pgo_profiles"))
-clean_files(profiles_dir)
+    if cleanup_needed:
+        print("\nCleaning up unneeded artifacts\n")
 
-thorium_dir = os.path.normpath(os.path.join(cr_src_dir, "out", "thorium"))
-delete_directory(thorium_dir)
+        files_deleted = (
+            clean_files(profiles_dir) if os.path.isdir(profiles_dir) else False
+        )
+        dir_deleted = (
+            delete_directory(thorium_dir) if os.path.exists(
+                thorium_dir) else False
+        )
 
-print("\nDone cleaning artifacts\n")
+        if files_deleted or dir_deleted:
+            print("\nDone cleaning unneeded artifacts\n")
+
+
+if __name__ == "__main__":
+    main()
